Sample data for camera trap events

A data set of invented camera trap event data.

Usage

event_data

Format

A tibble with 12 rows and 20 columns:

location_id

The ID code for the location of the camera trap

start_year

Year of start time of event

start_month

Month of start time of event

start_day

Day of start time of event

start_hour

Hour of start time of event

start_minute

Minute of start time of event

fa

Number of female adults

f1

Number of female yearlings

f0

Number of female calves

fu

Number of females of unknown age

ma

Number of male adults

m3

Number of male 3-year-olds

m2

Number of male 2-year-olds

m1

Number of male 1-year-olds

m0

Number of male calves

mu

Number of males of unknown age

ua

Number of adults of unknown sex

u1

Number of yearlings of unknown sex

u0

Number of calves of unknown sex

uu

Number of individuals of unknown age and unknown sex